requirements.txt
flask~=3.0.3
flask-pymongo
pymongo
flask-cors
werkzeug
gunicorn==21.2.0
flask_mail
jupyter~=1.0.0
pip~=24.0
panel~=1.4.4
attrs~=23.1.0
zlib~=1.2.13
distro~=1.9.0
wheel~=0.43.0
pysocks~=1.7.1
protobuf~=3.20.3
openssl~=3.0.15
pyopenssl~=24.0.0
cryptography~=42.0.5
pillow~=10.3.0
docutils~=0.18.1
sphinx~=7.3.7
tornado~=6.4.1
jinja2~=3.1.4
python-dotenv~=0.21.0